requirements.txt
streamlit
qdrant-client
sentence-transformers
python-dotenv
langchain-text-splitters
langchain-community
langchain-core
fastapi
google
protobuf
langchain-google-genai
SQLAlchemy
psycopg2-binary